+++ app/src/main/java/com/example/symptotrack/Doc/PatientItem.java
@@ -1,11 +1,11 @@
 package com.example.symptotrack.Doc;
 
 public class PatientItem {
-    public final int id;
+    public final long id;
     public final String nombreCompleto;
     public final String ultimoEnvio; // fecha texto
 
-    public PatientItem(int id, String nombreCompleto, String ultimoEnvio) {
+    public PatientItem(long id, String nombreCompleto, String ultimoEnvio) {
         this.id = id;
         this.nombreCompleto = nombreCompleto;
         this.ultimoEnvio = ultimoEnvio;
